Matrix Interpolation
Interpolates matrices from matrix A to matrix B linearly.
If used with value slider number goes from 0.0 to 1.0 and creates the
matrix between.
A list of floats can be used for factor input, then it generates the
matrices between A and B.
According to documentation values between 0.0 and 1.0 should work, but
larger values seem to work as well.