Update plane node centering to work correctly with step lists inputs
Previously the centering in the plane node did not take into account that the step size can come from a list hence it did not center correctly when a variable size step list was provided.

To center the grid correctly the size of the grid along X and Y is first calculated based on the given step lists and the vertex numbers.