Remove extra list nesting in torus knots node outputs
The previous implementation did not work well with the skin mesher node, which i suspect was due to the TK output lists having an extra nesting level. I removed it.

Note: The code was originally written like this, but as I tried to figure out how make the vectorization work well with torus knots that are degenerate (i.e. when p and q are not co-prime) I ended up created an extra nesting level. In the case  p and q are not co-prime the torus knot has multiple disjoint links which are part of the same torus knot. Since these links are separate lists in the output they will be replicated as independent units when using vectorization. The previous code was an attempt but it did not fix this problem anyway, so I chose to revert it to the first implementation so that the TK output works well with other nodes (e.g. like the skin mesher). For the time being this is a limitation until we figure out how to format the output so that the degenerate knot links are grouped together for vectorization.