Add experimental fully parametrized bounce easing function
Add fully parametrized bounce easing function to let user adjust number of bounces & attenuation. The derived formula is based on geometric progression series combined with a parabolic function for each bounce (see code comments)

Note: this is likely to be somewhat slower than the cryptic, simplistic version with hardcoded version from before, but it's comparable in speed, with the benefit of being fully parametrized.