Optimize easing functions via pre calcualted derived settings
Certain derived values within the easing functions dependent on the main parameters and do not change with p and thus can be cached before calling these functions for all values of p. This update introduces the prepare*Settings and default*Settings to be used to pre calculate these parameters.

The default values ensure compatibility with nodes using the easing functions relying on default behavior.

- update MixNumber node to use the optimized easing function version

- cleanup mixNumber node to remove old NA easing functions