Add custom icon support to nodeview space menu script - WIP
In order to show the icons in the search menu the script was updated to read the custom node's bl_icon2 attribute so that each node specifying a custom icon ID can have this custom icon displayed in the search menu.

Note: If the icon has a bl_icon field defined, this is going to be used for the menu icon, unless it has a bl_icon2 filed defined, in which case it will use the custom icon instead.