Add custom icon support
Add ui/sv_icons.py script to aid with the custom icons and a ui/icons folder with some testing icons (*.png).

Note: blender icons are 32px x 32px png format.

sv_icons functions:
* loadCustomIcons()
This will be called once during the registration of this module and it will load the custom icons from the ui/icons folder. The proposed filenames format for the custom icons is : sv_icon_name.png. The loading function reads all the png files in this folder matching this particular format and converts the filenames to upper case to be used as the icon string ID, which are used to reference the icons.

Example:

ui/icons/sv_torus_knot.png
will have a "SV_TORUS_KNOT" icon string ID created for it.

Note: alternatively we could create a collection of predefined icon string IDs and their corresponding icon filenames, though with this we would always have to touch the icons.py code whenever we add a new icon to the SV custom icon collection.

* removeCustomIcons()
This will unregister/remove the loaded custom icons.

* customIcon(name)
This will return the integer ID of a custom icon given its icon string ID.
If a custom icon does not exist for the given icon string ID, the function returns 0.

- register the script in the root __init__