Add slice scale input to Cylinder node generator
Add vectorized input socket to allow to scale each slice of the cylinder.
The default value is 1.0. If value is changed in place all slices are scaled by this factor. If float input list is given to the socket, each slice is scaled by the corresponding scale value.

The subdivision number dictates the clamping or the extension through repetition of this list (similar to how Nx and step_x work in line and plane).