Add flag to auto load icons if custom_icon called before icons are loaded
For the cases in which custom_icon is called before the sv_icons is registered (and hence the icons loaded) a flag was added to auto load the icons first time the custom_icon is called. Such an example is the case when item lists defined for a enumProperty which is to have custom items is defined outside of the class, and this list is initialized before the sv_icons module is registered.

Note: I added global again, same as for the icon dictionary until I figure out whether global is needed or not.