Add new Torus 2D node to the generator (extended) arsenal
This node will basically generate a 2D ring (aka 2D torus) alowing for the same kind of settings that the torus 3D offers, albeing adapted to the 2D space.

Minor/Major or Exterior/Interior radii could be adjusted to control the size and the width of the torus (ring) as well as the radial and circular sections. The radial sections can also be phased.

The node is pretty much a strip down version of the Torus 3D node so it is fully vectorized.

Note: The property list was compacted a little more (compared to Torus 3D node) to appease zeffii's quest to squash the vertical scroll as much as possible without compromising readability or stepping on the pep8 nerves ;)