Add expander flags to the SocketCommon
These are to be used with multirow input sockets to allow displaying a expand/collapse button to expand/collapse the input socket property.

Note: For now the use_expander flag is on for all nodes. This can be made False and let the individual nodes decide whether they wan their sockets to have an expander (via. self.inputs["name"].use_expander = True)