Add spin twist option to the torus node
The "spin twist" feature allows to rotate the spin sections of the torus such that the edges/polys appear as a continuous twist around the torus tube. The smallest spin twist is the "spin section delta angle" (= 2*pi / # spin sections). For this angle, for a complete revolution the last spin section is rotated relative to the first spin section by an angle as to shift the vertices in the spin section list by one vertex (essentially, one polygon shift in the spin direction). The total spin is typically an integer number of this smallest spin twist angle, which means in order to have the torus make a 2*pi (full) twist for a complete revolution, the "spin twist" value has to be equal to the number of spin sections T = Ns. For double full twists N = 2Ns etc.