Dev IO - Socket Custom Props Import/Export (WIP) (#1845)
* Return zero vector for no input in VerticesSocket

* Add Sv prefix to Color and Quaternion socket classes

* Add "link matrix" UI/operator for MatrixSocket

This is to allow the matrix socket to spawn a linked matrix-in node.

* Add "link vector" UI/operator for VerticesSocket

This is to allow the vertices socket to spawn a linked vector-in node.

* Add quaternion to matrix socket conversion

* Update collapse expander for color to use SvColorSocket name

* Add matrix to quaternion socket conversion

* Factorize the link new node code

Add draw_link_new_node to socket base (common) to allow various inputs to add button to spawn and link different input nodes.

* Minor class/idname changes to the link new node

* Minor edit

* Add "use quicklink" flag

To allow the quick link feature to be turned on/off if needed.

Note: the SV settings could allow for the sockets use_expander, expanded and use_quicklink to be set to ON/OFF.

* Tempo commit to use settings for use expander

* Some updtes to the socket conversion

* More socket updates

* Some other changes

* Remove preference settings for socket expander

    Keep things simple for now. Let the expander default behavior keep the nodes the same (with the sockets collapsed).

    In the future we may want to add options to preferences to let the user decide whether to cerate nodes with sockets expanded or not.

* Remove commented out obsolete draw methods

The socket specific draw calls were replaced by the draw call in the base class.

* Minor cleanup

* Comment out some prints

* Minor updates

- adjust the quicklink new node offset to avoid overlapping new nodes (new nodes are now offset based on their index).

* Add code to handle export/import of socket custom props

* Minor additions

Add use_custom_color to the superficial props.