Update 2ptspline node.
The code seems to be pre-prop_name-times. So in fact setting inputs as
properties was not supported.

This does not change algorithm; so the node still does not try to match
inputs data, I only added an exception instead of silently-not-working
in case when number of items does not match.

refs #1847