Add new Ellipse generator (extended) node
To compliment the trusty "circle" generator node.. here's a more versatile ellipse node, which makes the elliptical realm a bit more natural than fiddling with the circle node. :)

The documentation explains the gory details of the node.

In summary the ellipse can be defined either by the major/minor radii, by the major radius and the eccentricity or by the major radius and the focal length (represented by a "mode" setting). Additional parameters are given to further adjust the ellipse (phase, rotation and scale) as well as to center the ellipse around the its center and focal points.