Add feature to display all the SV nodes in a given category
The new feature is implemented as a panel, available in the Property Panel whenever the "Developer Mode" is enabled in SV preferences and it allows to display the SV nodes in a given category.

note: The code keeps track of the previously spawned nodes and removes them when nodes for a new category are being generated.

User can select a category from a drop-down (enum property) as well as navigate prev/next to cycle through the available categories.

A 1D bin-packing algorithm is used to compact the nodes in a given number of bins based on the user selected grid width, height or aspect ratio.

Inputs available to alter the node display:
- grid width/height
- grid X/Y spacing
- constrain layout (width,height,aspect)