Optimize the linear edge list generation using caching
Profiling the edge generation indicated that it is a slow operation and since this edge list is generated over and over in various nodes it makes sense to cache it. Profiling of various nodes (generating edges) indicated a speedup of almost two orders of magnitude in generating the edges compared to the list comprehension counterpart (e.g. edges = [[i, i+1] for i in range(N)].

With the proposed optimization, the list of edges are stored in an edge cache (e.g. [[0,1], [1,2]... [n-1,n]] .. and the cache is extended as longer edge lists are requested/generated. Any call to get_edge_list will return a subset of this edge cache thus not having to re-generate the same list over and over.

Various nodes like the line, spiral, torus knot, ellipse etc, which that generate lines with linear list of edges can benefit substantially from this speedup.

To get the linear edge list use:
edges = get_edge_list(n)
returning: [[0, 1], [1, 2], ... , [n-1, n]]

To get the loop edge list use:
edges = get_edge_loop(n)
returning: [[0, 1], [1, 2], ... , [n-2, n-1], [n-1, 0]]