Add Scalar-Vector mode to quaternion IN/OUT nodes
Scalar-Vector is the same as WXYZ except that the input/output is provided as scalar W + vector (X,Y,Z). This is to help with the work-flow when quaternions are regarded as a (s,v) pair.

note: The Scalar-Vector is different thant Axis-Angle mode as for this mode the quaternion is Q = cos(a/2) + sin(a/2) * vector(x,y,z)