Add (new) MeshSwitch node
The node allows to connect multiple mesh inputs (verts/edges/polys) and to select one of them to output. Useful for changing the input mesh model to be processed by a node tree.

New input sockets are added as new meshes are connected to allow the next mesh (verts/edges/polys) to be connected.