Inset Special node: possibility to work with inset = zero.
This adds a new parameter - "Zero inset faces". It defines what to do
with faces when inset is set to zero. Old behaivour (skip such faces as
if they were masked out) is defined by "Skip" mode. In "Fan" mode, such
faces are processed as usual - the internal face just becomes
degenerated and so is merged into single vertex, so we have a fan-like
stucture.