Add angle helper class to be used by nodes with angle parameters
Factorizing the angle units / conversion code into a helper class so that nodes that use angles that want to select the angle units (RAD, DEG, UNI) can have a way to create the the angle unit UI elements and conversions performed.

Add default setting to the SV settings for auto angle conversion when the units are changed.