Update nesting.rst
These changes are, more or less, a conclusion of what's been learned before, that I think help understand the info on the page. I'm not absolutely sure they are right, but I do think we need them as an introduction to the comprehensive info that comes next. It helps get the overall picture without exhaustive info.

It's really fundamental knowledge and one tends to get lost on these concepts here, as well as on lesson 02.

I think the need for having 3 brackets as input for viz nodes in lesson 02 needs better explanation and that explanation can be either there on Lesson 02 or here in this page as the final example is a specific example of a single face, and I found it hard to understand at first without comparing it with a regular example side by side (I might still haven't got it right though). I will deal with that later.