Fix: Correct node class mixin order to resolve poll function (#4686)
Python mixin order goes from left to right here, so when looking
for the node's poll function, it first finds a `poll` function from
`bpy.types.Node`, then Sverchok's base node class poll.

In https://developer.blender.org/T101259, this incorrect poll
caused Sverchok nodes to be visible in the Geometry Nodes editor.
While that would be fun, it doesn't work unfortunately. It is possible
to fix that too, but the polls should be fixed anyway.

An alternative solution would be to somehow avoid defining the poll
returning "True" in `bpy.types.Node`, but I'm not quite sure how that
would work, and it might have more unexpected consequences.