Menu presets: automatic application.
Add a possibility to apply the selected menu preset automatically, at
each blender startup.

This is aimed for cases when we added a new node (or just rearranged
something in the menu preset), but user's index.yaml in datafiles still
needs to be updated.

If the feature is not enabled, it means the user takes the
responsibility on updating his index.yaml manually.